+++ backend/src/main/java/project/backend/domain/member/app/MemberService.java
@@ -8,7 +8,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import project.backend.domain.imagefile.ImageFile;
 import project.backend.domain.imagefile.ImageFileService;
-import project.backend.domain.member.MemberErrorCode;
 import project.backend.domain.member.dao.MemberRepository;
 import project.backend.domain.member.dto.MemberResponse;
 import project.backend.domain.member.dto.MemberUpdateRequest;
@@ -16,6 +15,7 @@ import project.backend.domain.member.dto.SignUpRequest;
 import project.backend.domain.member.entity.Member;
 import project.backend.domain.member.mapper.MemberMapper;
 import project.backend.global.exception.MemberException;
+import project.backend.global.exception.errorcode.MemberErrorCode;
 
 @Slf4j
 @Service
