--- backend/src/main/java/project/backend/domain/member/dto/ErrorResponse.java
@@ -1,17 +1,0 @@
-package project.backend.domain.member.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ErrorResponse {
-
-    private final String code;
-    private final String message;
-    private HttpStatus status;
-
-}
