+++ backend/src/main/java/project/backend/global/exception/ex/BaseException.java
@@ -1,8 +1,10 @@
 package project.backend.global.exception.ex;
 
+import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import project.backend.global.exception.errorcode.ErrorCode;
 
+@Getter
 public abstract class BaseException extends RuntimeException {
 
     private final ErrorCode errorCode;
@@ -10,10 +12,6 @@ public abstract class BaseException extends RuntimeException {
     public BaseException(ErrorCode errorCode) {
         super(errorCode.getMessage());
         this.errorCode = errorCode;
-    }
-
-    public ErrorCode getErrorCode() {
-        return errorCode;
     }
 
     public HttpStatus getStatus() {
