+++ frontend/src/pages/Home.js
@@ -7,17 +7,15 @@ const Home = () => {
   const [loading, setLoading] = useState(true); // 로딩 상태
 
   useEffect(() => {
-    const token = localStorage.getItem('accessToken'); // 또는 sessionStorage
+    // const token = localStorage.getItem('accessToken'); // 또는 sessionStorage
 
-    if (!token) {
-      navigate('/login'); // 토큰 없으면 로그인
-      return;
-    }
+    // if (!token) {
+    //   navigate('/login'); // 토큰 없으면 로그인
+    //   return;
+    // }
 
     axios.get(`http://localhost:8080/chat-rooms/recent`, { 
-        headers: {
-            Authorization: `Bearer ${token}` // ✅ JWT 토큰을 헤더에 첨부
-        }
+        withCredentials: true,
     })
       .then(res => {
         const roomId = res.data.roomId;
