+++ frontend/src/pages/login-form.jsx
@@ -2,43 +2,43 @@
 
 import { useState } from "react"
 import { useNavigate } from "react-router-dom";
-import "./App.css"
+import "../App.css"
 
 function App() {
   const navigate = useNavigate();
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
-  const [nickname, setNickname] = useState("")
 
   const handleSubmit = async (e) => {
   e.preventDefault();
 
-  const payload = {
-    email,
-    password,
-    nickname,
-  };
+  const params = new URLSearchParams();
+  params.append("email",email);
+  params.append("password",password);
 
   try {
-    const response = await fetch("http://localhost:8080/signup", {
+    const response = await fetch("http://localhost:8080/login", {
       method: "POST",
       headers: {
-        "Content-Type": "application/json",
+        "Content-Type": "application/x-www-form-urlencoded",
+        "Accept": "application/json",
       },
-      body: JSON.stringify(payload),
+      body: params.toString(),
+      credentials: "include",
     });
+
 
     if (!response.ok) {
       const errorData = await response.json();
       console.error(errorData)
-      alert(errorData.message || "회원가입 실패");
+      alert(errorData.message || "로그인 실패");
       return;
     }
 
     const data = await response.json();
-    console.log("✅ Signup successful:", data);
-    alert("회원가입 성공!");
-    navigate("/login");
+    console.log("Signup successful:", data);
+    alert("로그인 성공!");
+    navigate("/");
 
   } catch (error) {
     console.error("네트워크 또는 서버 에러:", error);
@@ -55,7 +55,7 @@ function App() {
           </a>
         </div>
 
-        <h1 className="heading">Let's Start!</h1>
+        <h1 className="heading">Welcome Back!</h1>
 
         <form onSubmit={handleSubmit}>
           <div className="form-group">
@@ -82,26 +82,14 @@ function App() {
             />
           </div>
 
-          <div className="form-group">
-            <label htmlFor="nickname">Username</label>
-            <input
-              id="nickname"
-              type="text"
-              value={nickname}
-              onChange={(e) => setNickname(e.target.value)}
-              placeholder="사용할 닉네임을 입력해주세요"
-              required
-            />
-          </div>
-
           <button type="submit" className="signup-button">
-            Sign Up
+            Login
           </button>
         </form>
 
         <div className="signup-link">
-          <span>이미 계정이 있나요?</span>
-          <a href="/login">로그인</a>
+          <span>계정이 아직 없으신가요?</span>
+          <a href="/signup">회원가입</a>
         </div>
       </div>
