+++ frontend/src/pages/signup.jsx
@@ -2,43 +2,43 @@
 
 import { useState } from "react"
 import { useNavigate } from "react-router-dom";
-import "./App.css"
+import "../App.css"
 
 function App() {
   const navigate = useNavigate();
   const [email, setEmail] = useState("")
   const [password, setPassword] = useState("")
+  const [nickname, setNickname] = useState("")
 
   const handleSubmit = async (e) => {
   e.preventDefault();
 
-  const params = new URLSearchParams();
-  params.append("email",email);
-  params.append("password",password);
+  const payload = {
+    email,
+    password,
+    nickname,
+  };
 
   try {
-    const response = await fetch("http://localhost:8080/login", {
+    const response = await fetch("http://localhost:8080/signup", {
       method: "POST",
       headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        "Accept": "application/json",
+        "Content-Type": "application/json",
       },
-      body: params.toString(),
-      credentials: "include",
+      body: JSON.stringify(payload),
     });
-
 
     if (!response.ok) {
       const errorData = await response.json();
       console.error(errorData)
-      alert(errorData.message || "로그인 실패");
+      alert(errorData.message || "회원가입 실패");
       return;
     }
 
     const data = await response.json();
-    console.log("Signup successful:", data);
-    alert("로그인 성공!");
-    navigate("/chat");
+    console.log("✅ Signup successful:", data);
+    alert("회원가입 성공!");
+    navigate("/login");
 
   } catch (error) {
     console.error("네트워크 또는 서버 에러:", error);
@@ -55,7 +55,7 @@ function App() {
           </a>
         </div>
 
-        <h1 className="heading">Welcome Back!</h1>
+        <h1 className="heading">Let's Start!</h1>
 
         <form onSubmit={handleSubmit}>
           <div className="form-group">
@@ -82,14 +82,26 @@ function App() {
             />
           </div>
 
+          <div className="form-group">
+            <label htmlFor="nickname">Username</label>
+            <input
+              id="nickname"
+              type="text"
+              value={nickname}
+              onChange={(e) => setNickname(e.target.value)}
+              placeholder="사용할 닉네임을 입력해주세요"
+              required
+            />
+          </div>
+
           <button type="submit" className="signup-button">
-            Login
+            Sign Up
           </button>
         </form>
 
         <div className="signup-link">
-          <span>계정이 아직 없으신가요?</span>
-          <a href="/signup">회원가입</a>
+          <span>이미 계정이 있나요?</span>
+          <a href="/login">로그인</a>
         </div>
       </div>
